Collision detection semi-working. Algorithm needs to handle all three directions of motion simultaneously, not separately. Next up is block placement/destruction.

+++ Terrain.h
@@ -22,8 +22,8 @@ public:
     
     void draw();
     
-    GLuint getBlock(int x, int y, int z);
-    void setBlock(int x, int y, int z,GLubyte blockID);
+    GLuint getBlock(double x, double y, double z);
+    void setBlock(double x, double y, double z,GLubyte blockID);
     std::vector<std::vector<std::vector<Chunk> > > chunks;
     int originX,originY,originZ;
 };
